+++ frontend/src/components/notifications/NotificationsTable.tsx
@@ -90,6 +90,7 @@ export default function NotificationsTable() {
                             <TableCell align="right">{t("RECIPIENT_NAME")}</TableCell>
                             <TableCell align="right">{t("RECIPIENT_EMAIL")}</TableCell>
                             <TableCell align="right">{t("DATA")}</TableCell>
+                            <TableCell align="right">{t("ERROR")}</TableCell>
                         </TableRow>
                     </TableHead>
                     <TableBody>
@@ -115,6 +116,7 @@ export default function NotificationsTable() {
                                                                 state: "pending",
                                                                 num_sending_attempts: 0,
                                                                 last_sending_attempt_at: null,
+                                                                traceback: null,
                                                             },
                                                         })
                                                     );
@@ -147,6 +149,9 @@ export default function NotificationsTable() {
                                 </TableCell>
                                 <TableCell align="right">
                                     {JSON.stringify(notification.payload)}
+                                </TableCell>
+                                <TableCell align="right">
+                                    {notification.traceback}
                                 </TableCell>
                             </TableRow>
                         ))}