+++ server/src/notifications/types.py
@@ -1,4 +1,4 @@
-from typing import Optional
+from typing import Optional, List
 
 from pydantic import BaseModel
 
@@ -7,3 +7,4 @@ class NotificationPayloadSchema(BaseModel):
     flight_id: Optional[int] = None
     tow_pilot_id: Optional[int] = None
     tow_airplane_id: Optional[int] = None
+    flights_ids: Optional[List[int]] = None